refactor(config): extract provider validation boundary
